backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
kubernetes>=28.1.0
python-multipart>=0.0.6
httpx>=0.26.0
structlog>=24.1.0
# Auth dependencies
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
# PostgreSQL
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
# Migrations
alembic>=1.13.0
# Additional crypto (for kubeconfig encryption)
cryptography>=41.0.0
# AI/ML
openai>=1.10.0
tiktoken>=0.5.2

